refactor: move app-client code to it's own file
